fix: address Copilot review feedback on PR #2
- Add ModelName field to LlamaCppConfig for configurable model identifier
- Pass model config through registry and module factories for llama_cpp
- Pin llama-server download to specific version tag instead of latest
- Add User-Agent header to GitHub API requests
- Fix zombie process: Wait() after Kill() in Close() and ensureServer
- Use http.NewRequestWithContext for health check (cancelable)
- Handle url.Parse error in OllamaProvider constructor
- Sanitize HuggingFace filename to prevent path traversal
- Use int64 for size_bytes in step.model_pull output

+++ step_model_pull_test.go
@@ -115,7 +115,7 @@ func TestModelPullStep_HuggingFace(t *testing.T) {
 	if result.Output["model_path"] == "" {
 		t.Error("expected non-empty model_path")
 	}
-	if result.Output["size_bytes"].(int) != len(content) {
+	if result.Output["size_bytes"].(int64) != int64(len(content)) {
 		t.Errorf("size_bytes: want %d, got %v", len(content), result.Output["size_bytes"])
 	}
 }
